Rename Module files to follow convention
