+++ lib/game_logic.rb
