+++ lib/modules/game_text.rb
