+++ lib/modules/game_utils.rb
